+++ .eslintignore
@@ -1,3 +1,3 @@
 config
 tools
-components/base/Content/TreeView/zTreew
+components/base/Content/Tree/zTreew
