+++ src/components/box/GridView/grid-view.vue
@@ -54,9 +54,9 @@
             </b-button-group>
             <b-dropdown
               :list="downloadList"
-              menu-align="right"
               hide-toggle
-              @menuClick="dataExport"
+              menu-align="right"
+              @menu-click="item => dataExport(item)"
             >
               <template #trigger>
                 <b-button
@@ -88,14 +88,19 @@
           :class="fixedNum > 0 ? `col-${fixedSizeNum}` : ''"
           :list="fixedList"
           :sort-obj="sortObj"
-          :table-class="tableClass"
-          :thead-class="theadClass"
+          :table-theme="tableTheme"
+          :table-sm="tableSm"
+          :table-hover="tableHover"
+          :table-striped="tableStriped"
+          :table-bordered="tableBordered"
+          :table-borderless="tableBorderless"
+          :thead-theme="theadTheme"
           :hide-head="hideHead"
           :hide-data="hideData"
           :hide-foot="hideFoot"
           :hide-serial="hideSerial"
-          :primary-key="primaryKey"
           :select-status="selectStatus"
+          :primary-key="primaryKey"
           @table:sort="cell => tableSort(cell)"
           @table:scroll="(event, type) => scroll(event, type)"
         />
@@ -104,19 +109,24 @@
           v-if="fixedNum > 0"
           ref="activeTable"
           :list="activeList"
-          :class="`col-${12 - fixedSizeNum}`"
           is-active
           hide-serial
           hide-select
+          :class="`col-${12 - fixedSizeNum}`"
           :sort-obj="sortObj"
-          :table-class="tableClass"
-          :thead-class="theadClass"
+          :table-theme="tableTheme"
+          :table-sm="tableSm"
+          :table-hover="tableHover"
+          :table-striped="tableStriped"
+          :table-bordered="tableBordered"
+          :table-borderless="tableBorderless"
+          :thead-theme="theadTheme"
           :hide-head="hideHead"
           :hide-data="hideData"
           :hide-foot="hideFoot"
-          :primary-key="primaryKey"
           :selected="selectedOptions"
           :select-status="selectStatus"
+          :primary-key="primaryKey"
           @table:sort="cell => tableSort(cell)"
           @table:scroll="(event, type) => scroll(event, type)"
         />
@@ -149,7 +159,7 @@
             start="1"
             :end="pageCount"
           >
-            <b-number />
+            <b-number v-model.number="pageNumber" />
             <!-- <b-button class="mx-1" size="sm" value="跳转" outline /> -->
           </b-pagination>
         </font>
@@ -194,8 +204,8 @@
           </font>
           <div class="col-9">
             <b-select
-              size="sm"
               :list="['asc', 'desc']"
+              size="sm"
               :value="sortPlusObj[item]"
               @change="sortPlusChanged($event, item)"
             />
@@ -230,13 +240,14 @@
 </template>
 
 <script>
-import tools from "@/tools/index.js";
-import config from "@/config/index.js";
-import util from "@/components/util/index.js";
 // 参考： https://printjs.crabbly.com/
 import printJS from "print-js";
 
-import CTable from "./c-table.vue";
+import tools from "@/tools/index.js";
+import config from "@/config/index.js";
+import util from "@/components/util/index.js";
+
+import CTable from "@/components/content/Table/c-table.vue";
 
 import BButton from "@/components/basic/Button/basic-button.vue";
 import BButtonGroup from "@/components/base/ButtonGroup/b-button-group.vue";
@@ -287,19 +298,20 @@ export default {
       validator: value => !isNaN(value) && [0, 1, 2].includes(Number(value))
     },// 0: 默认, 1: 单选, 2: 多选
     selected: [Array, Object],
-    printTitle: String
+    printTitle: util.props.String
   },
   data() {
     return {
       loading: false, // 未使用
       selectedOptions: this.selected,
       downloadList: [
-        { value: "JSON", type: "json" },
         { value: "XML", type: "xml" },
         { value: "CSV", type: "csv" },
         { value: "TXT", type: "txt" },
-        { value: "SQL", type: "sql" },
-        { value: "MS-EXCEL", type: "ms-excel" }
+        // { value: 'SQL', type: 'sql', },
+        // { value: 'PDF', type: 'pdf', },
+        { value: "JSON", type: "json" }
+        // { value: 'MS-EXCEL', type: 'ms-excel', },
       ],
       sortObj: {},
       sortPlusObj: {},
@@ -389,18 +401,6 @@ export default {
     },
     hideFoot: function() {
       return !this.foot || this.foot.length == 0;
-    },
-    tableClass: function() {
-      let theme = this.tableTheme ? `table-${this.tableTheme}` : "";
-      let sm = this.tableSm ? "table-sm" : "";
-      let hover = this.tableHover ? "table-hover" : "";
-      let striped = this.tableStriped ? "table-striped" : "";
-      let bordered = this.tableBordered ? "table-bordered" : "";
-      let borderless = this.tableBorderless ? "table-borderless" : "";
-      return `${theme} ${hover} ${striped} ${bordered} ${borderless} ${sm} `;
-    },
-    theadClass: function() {
-      return this.theadTheme ? `thead-${this.theadTheme}` : "";
     },
     fixedTable: function() {
       return this.$refs && this.$refs.fixedTable;
@@ -599,6 +599,32 @@ export default {
     },
     dataExport: function(item) {
       if (!item || !item.type) return;
+      switch (item.type) {
+        case "xml":
+          tools.file.xml.writer(this.data);
+          break;
+        case "csv":
+          tools.file.csv.writer(this.data);
+          break;
+        case "txt":
+          tools.file.txt.writer(this.data);
+          break;
+        // case 'sql':
+
+        //     break;
+        // case 'pdf':
+
+        //     break;
+        case "json":
+          tools.file.json.writer(this.data);
+          break;
+        case "ms-excel":
+          tools.file.excel.writer(this.data);
+          break;
+        default:
+          break;
+      }
+      return false;
     },
     tableSort: function(cell) {
       this.$set(
