+++ src/components/box/VideoPlayer/Plugins/Playlist/ListCard/list-card.vue
