+++ src/components/box/VideoPlayer/Plugins/Playlist/Selector/selector.vue
