+++ src/components/box/VideoPlayer/Plugins/Playlist/list/list.vue
