+++ src/components/content/Table_1/Foot/table-foot.vue
