+++ src/components/content/Table_1/c-table_1.vue
